% Converting variable type

+++ app/src/main/java/com/dvidal/samplereviews/features/list/data/local/config/ConfigDto.kt
@@ -13,6 +13,6 @@ data class ConfigDto(
     val activityName: String = "",
     val offsetPage: Int = 0,
     val numReviews: Int = 0,
-    val averageRating: Float = 0.0F,
+    val averageRating: Double = 0.0,
     val isDescendingOrderRating: Boolean = true
 )
